Thêm sửa xóa khách hàng
Thêm sửa xóa khách hàng

+++ src/ClockDetail/Controller/ClockDetailController.java
@@ -143,7 +143,7 @@ public class ClockDetailController extends BaseController {
         } else {
             this.frmClockDetail.getLblCustomerName().setText("");
         }
-        int customerID = customer.getId();
+        long customerID = customer.getId();
        
 
         try {
